Review of repository class structure
* JournalData is not a repository. It doesn't work directly with DB
tables and it contains "business logic". This makes it a controller.
Therefore I moved it to controller module.
* DB initialization is not a typical repository responsibility. It's
better to isolate this functionality to a dedicated function (or class).
* Repositories don't typically establish their own connection to a DB.
It's given to them by a 3rd party. You're not dealing with multiuser,
highly parallel system here, hence a single connection passed to the
repository upon its construction should suffice.
